Working toward modularized webpack config

+++ cookiecutter/.cookiecutters/js/node/hooks/post_gen_project.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
 # Pull node dependencies into package.json
+if [ '{{cookiecutter.dependencies}}' ]
+then
 
-cp package.json package-base.json
+    cp package.json package-base.json
 
-cd "{{cookiecutter.dependency_dir}}"
+    cd "{{cookiecutter.dependency_dir}}"
 
-cat "$OLDPWD/package-base.json" `find  -L {{cookiecutter.dependencies}} -iname "package.json"` | json --deep-merge > "$OLDPWD/package.json"
+    cat "$OLDPWD/package-base.json" `find  -L {{cookiecutter.dependencies}} -iname "package.json"` | json --deep-merge > "$OLDPWD/package.json"
 
-cd "$OLDPWD"
+    cd "$OLDPWD"
 
-rm package-base.json
+    rm package-base.json
+
+fi
